move protos to root
