+++ tsconfig.json
@@ -7,7 +7,6 @@
   "include": [
     "**/*.ts",
     ".eslintrc.js",
-    "integration_test/**/*",
-    "src/protos/**/*"
+    "integration_test/**/*"
   ]
 }
