+++ tsconfig.release.json
@@ -16,7 +16,6 @@
     "./src/logger/index.ts",
     "./src/logger/compat.ts",
     "./src/v2/index.ts",
-    "./src/bin/firebase-functions.ts",
-    "./src/protos/index.ts",
+    "./src/bin/firebase-functions.ts"
   ]
 }
